Manifest comparator pattern implemented for S3
