+++ gradle.properties
@@ -1,4 +1,4 @@
-version=0.85
+version=0.85.2
 springVersion=4.2.6.RELEASE
 springBootVersion=1.3.5.RELEASE
 jerseyVersion=2.22.2
