--- micro-couchbase/src/main/java/com/aol/micro/server/couchbase/base/Data.java
@@ -1,23 +1,0 @@
-package com.aol.micro.server.couchbase.base;
-
-import java.io.Serializable;
-import java.util.Date;
-
-import lombok.Getter;
-
-public class Data<T> implements Serializable{
-	
-	private static final long serialVersionUID = 1L;
-	@Getter
-	private final T data;
-	@Getter
-	private final Date date;
-	@Getter
-	private  final String versionedKey;
-	
-	public Data(T data, Date date, String versionedKey){
-		this.data = data;
-		this.date = date;
-		this.versionedKey = versionedKey;
-	}
-}
