--- micro-couchbase/src/main/java/com/aol/micro/server/couchbase/base/ManifestComparatorKeyNotFoundException.java
@@ -1,12 +1,0 @@
-package com.aol.micro.server.couchbase.base;
-
-
-
-public class ManifestComparatorKeyNotFoundException extends RuntimeException{
-
-	public ManifestComparatorKeyNotFoundException(String message) {
-		super(message);
-		
-	}
-
-}
