+++ micro-manifest-comparator/build.gradle
@@ -5,19 +5,13 @@ repositories {
 	maven { url "https://jitpack.io" }
 }
 
-description = 'micro-couchbase'
+description = 'micro-manifest-comparator'
 dependencies {
-	compile group: 'com.couchbase.client', name: 'couchbase-client', version:'1.4.8'
-	compile project(':micro-manifest-comparator')
+	
 	compile project(':micro-core')
-	compile project(':micro-guava')
-	testCompile group: 'org.codehaus.groovy', name: 'groovy-all', version:'2.3.3'
-	testCompile(group: 'org.spockframework', name: 'spock-core', version:'0.7-groovy-2.0') { exclude(module: 'groovy-all') }
-	testCompile group: 'com.cyrusinnovation', name: 'mockito-groovy-support', version:'1.3'
-	testCompile 'cglib:cglib-nodep:2.2'
-	testCompile  'com.github.johnmcclean-aol:couchbasemock:master'
+	compile project(':micro-jackson-configuration')
 	testCompile group: 'org.hamcrest', name: 'hamcrest-all', version:hamcrestVersion
-	testCompile project(':micro-grizzly-with-jersey')
+	
 }
 
 sourceSets {
@@ -36,8 +30,8 @@ sourceSets {
 
 modifyPom {
 	project {
-		name 'Microserver couchbase'
-		description 'Opinionated rest microservices'
+		name 'Microserver Manifest Comparator'
+		description 'Pattern for loading cached data for BASE microservices'
 		url 'https://github.com/aol/micro-server'
 		inceptionYear '2015'
