+++ micro-manifest-comparator/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Jun 28 15:05:52 IST 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.13-bin.zip
