+++ micro-manifest-comparator/src/main/java/com/aol/micro/server/manifest/ManifestComparator.java
@@ -1,0 +1,11 @@
+package com.aol.micro.server.manifest;
+
+public interface ManifestComparator<T> {
+	public <R> ManifestComparator<R> withKey(String key);
+	public boolean isOutOfDate();
+	public void load() ;
+	public void cleanAll();
+	public void clean(int numberToClean);
+	public void saveAndIncrement(T data);
+	public T getData();
+}
