+++ micro-manifest-comparator/src/main/java/com/aol/micro/server/manifest/ManifestComparatorKeyNotFoundException.java
@@ -1,0 +1,14 @@
+package com.aol.micro.server.manifest;
+
+
+
+public class ManifestComparatorKeyNotFoundException extends RuntimeException{
+
+	private static final long serialVersionUID = 1L;
+
+	public ManifestComparatorKeyNotFoundException(String message) {
+		super(message);
+		
+	}
+
+}
