+++ micro-s3/src/main/java/com/aol/micro/server/s3/data/S3Deleter.java
@@ -1,0 +1,15 @@
+package com.aol.micro.server.s3.data;
+
+import com.amazonaws.services.s3.AmazonS3Client;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class S3Deleter {
+	private final String bucket;
+	private final AmazonS3Client client;
+	
+	public void delete(String key){
+		client.deleteObject(bucket, key);
+	}
+}
