+++ micro-s3/src/main/resources/META-INF/services/com.aol.micro.server.Plugin
@@ -1,1 +1,1 @@
-com.aol.micro.server.s3.S3Plugin
+com.aol.micro.server.s3.plugin.S3Plugin
