+++ settings.gradle
@@ -33,3 +33,4 @@ include ':micro-logback'
 include ':micro-s3'
 include ':micro-tutorial'
 include ':micro-dist-lock'
+include ':micro-manifest-comparator'
